Refactoring changes: datamart_initializer renamed to datamart_initializer_sqlite
