+++ inverted_index/datamart_initializer_sqlite.py
